backend/requirements.txt
flask==2.0.1
werkzeug==2.0.1
python-dotenv==0.19.0
google-generativeai==0.8.4
flask-cors==4.0.0
requests>=2.31.0